make explicit a previously-implicit qint64-to-size_t conversion [-Wconversion]

+++ src/helper/helper.cpp
@@ -182,7 +182,7 @@ public:
         size_ += n_bytes;
         sized_ += double(n_bytes);
 
-        history_.add(clock_(), n_bytes);
+        history_.add(clock_(), size_t(n_bytes));
 
         update_percent_done();
     }
